+++ Clicked/Definitions.lua
@@ -176,3 +176,9 @@
 
 --- @class BindingReloadCauses : BindingReloadCause
 --- @field public binding BindingReloadCause
+
+--- @class TalentInfo
+--- @field public entryId integer
+--- @field public spellId integer
+--- @field public text string
+--- @field public icon integer